feat: add flamegraph profiler macro
