+++ .gitignore
@@ -21,3 +21,6 @@ Cargo.lock
 # IDE
 .idea
 .vscode
+
+**/*.svg
+.DS_Store
