+++ profiler_rs/src/timer_profiler_test/docs_test.rs
