+++ profiler_rs/src/timer_profiler_test/trait_test.rs
